Move message commands to command package Change #getVersion to a local variable
