+++ src/main/java/me/noodles/scp/commands/messages/Store.java
@@ -1,4 +1,4 @@
-package me.noodles.scp.messagecommands;
+package me.noodles.scp.commands.messages;
 
 import me.noodles.scp.utilities.Common;
 import org.bukkit.command.CommandSender;
@@ -11,16 +11,16 @@ import me.noodles.scp.SCP;
 import java.util.Collections;
 import java.util.List;
 
-public final class Teamspeak implements TabExecutor {
+public final class Store implements TabExecutor {
 
 	@Override
 	public boolean onCommand(final CommandSender sender, final Command command, final String label, final String[] args) {
-		if (getPlugin().getConfig().getBoolean("Teamspeak.Enabled")) {
+		if (getPlugin().getConfig().getBoolean("Store.Enabled")) {
 			if (sender instanceof Player) {
 				final Player player = (Player) sender;
 
-				if (player.hasPermission("scp.teamspeak")) {
-					Common.tell(player, getPlugin().getConfig().getString("Messages.TeamspeakMessage"));
+				if (player.hasPermission("scp.store")) {
+					Common.tell(player, getPlugin().getConfig().getString("Messages.StoreMessage"));
 
 					return true;
 				}
